+++ test/utests/fakes/FakeAampMPDUtils.cpp
@@ -66,3 +66,68 @@ bool ParseSegmentIndexBox( const char *start, size_t size, int segmentIndex, uns
 {
 	return true;
 }
+
+/**
+ * @brief Read 16 word helper function
+ * @param pptr pointer to read from
+ * @retval word value
+ */
+unsigned int Read16( const char **pptr)
+{
+	return 0;
+}
+
+/**
+ * @brief Read 32 word helper function
+ * @param pptr pointer to read from
+ * @retval word value
+ */
+unsigned int Read32( const char **pptr)
+{
+	return 0;
+}
+
+/**
+ * @brief Read 64 word helper function
+ * @param pptr pointer to read from
+ * @retval word value
+ */
+uint64_t Read64( const char **pptr)
+{
+	return 0;
+}
+
+/**
+ * @brief read unsigned multi-byte value and update buffer pointer
+ * @param[in] pptr buffer
+ * @param[in] n word size in bytes
+ * @retval 32 bit value
+ */
+uint64_t ReadWordHelper( const char **pptr, int n )
+{
+	return 0;
+}
+
+/**
+ * @brief Replace matching token with given number
+ * @param str String in which operation to be performed
+ * @param from token
+ * @param toNumber number to replace token
+ * @retval position
+ */
+int replace(std::string &str, const std::string &from, uint64_t toNumber)
+{
+	return 0;
+}
+
+/**
+ * @brief Replace matching token with given string
+ * @param str String in which operation to be performed
+ * @param from token
+ * @param toString string to replace token
+ * @retval position
+ */
+int replace(std::string &str, const std::string &from, const std::string &toString)
+{
+	return 0;
+}
